requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
pytest==8.3.5
pytest-asyncio==0.25.3
